+++ Changelog.md
@@ -1,4 +1,4 @@
-Continuous-Game-Play.py:
+/////////////////Continuous-Game-Play.py/////////////////
 
 v1.0:
 - This is the first version
@@ -14,7 +14,7 @@ v1.2:
 -Pushed (this) changelog from Replit
 -Fixed errors in Python Style Guide
 
-RPG-Inventory.py:
+////////////////////RPG-Inventory.py////////////////////
 
 v1.0:
 
@@ -36,3 +36,35 @@ v1.4:
 
 - Made changes to follow the python style guide
 - Bug fixes
+
+/////////////////RPG-Modules-and-Maps.py/////////////////
+
+v1.0:
+
+- Separated RPG-Inventory.py into different modules
+- This is the first version
+
+v1.1:
+
+- Added Maps.py
+- Bug fixes
+- Defined player location
+
+v1.2:
+
+- Added alternate lists to insert into map to show player's location
+- Removed circular imports
+
+v1.3:
+- Bug fixes related to the game map
+- Attempted to remove duplicate "You're here" tiles on map
+
+v1.4:
+- Removed duplicate "You're here" tiles 
+- Made the Healing sword not available in market until the rusty knife is purchased.
+
+v1.5:
+- Added purchase functions with paramaters to make buy_items() neater
+- Bug fixes
+- Checked PEP8
+- Fixed recursion issue